Improve Usability (#1)
* add `const` to keyword list

Adding const to the keyword list results in it being higlighted correctly by
Emacs.

* add syntax table to typespec-ts-mode

Marks comments correctly, thus allowing users to use the
`prog-fill-reindent-defun` functionality. Without the syntax table, comments are
not determined as strings, thus the filling fails.

* use c-style comments

Other modes (typescript-ts-mode) also use this to setup their comments.
Typespec has the same comment style as Typescript.

* fix indentation to match the default Typespec code style

Typespec has a default code style. Previously, everything was indented by
`typespec-ts-mode-indent-offset`, except for some symbols. Requiring the code
formatter to reformat the code on trivial changes. Now most elements are
indented correctly, including comments. The indentation is implemented such that
the formatter does not necessary has to reformat the code.

The main motivation of this change was to fix the behavior where the following
was indented wrong (where | is the cursor):

  @doc("...")|
  model Test {

Would be indented the following, if inserting a new line (pressing enter):

  @doc("...")
    |
  model Test {

Another issue was:

  @doc("...")
  model Test {|

Was indented into

  @doc("...")
    model Test {
    |

This is fixed by specifying each element type separately.

* add defun definitions

Specifies elements which can be used for `defun` thing-at-point queries.